request to sign up properly sends and gets response from stand

+++ regsystem.parent/regsystem.controller/src/main/java/kz/zhadyrassyn/regsystem/controller/model/request/SignUpRequest.java
@@ -12,7 +12,7 @@ public class SignUpRequest {
     public String email;
     public String password;
     public String phone;
-    public int groupId;
+    public long groupId;
     public GenderEnum gender;
     public Date birthDate;
